Dubpob-matrix : Implenting calculation of Inverse matrix with Adjudent method.

+++ dubpob-matrix/src/main/java/org/dubpob/matrix/Matrix.java
@@ -5,7 +5,7 @@ public abstract class Matrix {
 	protected int width = 0;
 	
 	public abstract String getTrace();
-	public abstract int getDeterminant();
+	public abstract float getDeterminant();
 	@Override
 	public abstract String toString();
