+++ dubpob-matrix/src/main/java/org/dubpob/matrix/impl/SimpleMatrix.java
@@ -4,38 +4,38 @@ import org.dubpob.matrix.Matrix;
 import org.dubpob.matrix.utils.MatrixHelper;
 
 public class SimpleMatrix extends Matrix {
-	private int[][] matrix = null;
+	private float[][] matrix = null;
 	
 	
 	public SimpleMatrix(int height,int width) {
-		matrix = new int[height][width];
+		matrix = new float[height][width];
 		this.height = height;
 		this.width = width;
 	}
 
-	public int[][] getMatrix() {
+	public float[][] getMatrix() {
 		return matrix;
 	}
 
-	public void setMatrix(int[][] matrix) {
+	public void setMatrix(float[][] matrix) {
 		this.matrix = matrix;
 	}
 	
 	@Override
 	public String getTrace() {
 		if(height == width) {
-			int trace = 0;
+			float trace = 0;
 			for(int i = 0; i < height; i++) {
 				trace += matrix[i][i];
 			}
-			return Integer.toString(trace);
+			return Float.toString(trace);
 		}
 		return "";
 	}
 
 	@Override
-	public int getDeterminant() {
-		int value = 0;
+	public float getDeterminant() {
+		float value = 0;
 		if(width == height && matrix != null) {
 			switch(width) {
 				case 1:
